LC-third-maximum-number [Runtime: N/A; Memory: N/A]

+++ sorting/third-maximum-number.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int thirdMax(vector<int>& nums) {
 
-        int largest = INT_MIN;
-        int second = INT_MIN;
-        int third = INT_MIN;
+        long long largest = LLONG_MIN;
+        long long second = LLONG_MIN;
+        long long third = LLONG_MIN;
 
         for(int x : nums){
             if(x > largest){
@@ -23,7 +23,7 @@ public:
                 third = x;
             }
         }
-        if(third == INT_MIN){
+        if(third == LLONG_MIN){
             return largest;
         }
